# perf(napi): misc napi optimizations