--- a/ext/napi/util.rs
+++ b/ext/napi/util.rs
@@ -118,18 +118,18 @@ pub(crate) unsafe fn check_new_from_utf8_len<'s>(
   );
   return_error_status_if_false!(env, !str_.is_null(), napi_invalid_arg);
   let string = if len == NAPI_AUTO_LENGTH {
-    let result = unsafe { std::ffi::CStr::from_ptr(str_ as *const _) }.to_str();
-    return_error_status_if_false!(env, result.is_ok(), napi_generic_failure);
-    result.unwrap()
+    unsafe { std::ffi::CStr::from_ptr(str_ as *const _) }.to_bytes()
   } else {
     let string = unsafe { std::slice::from_raw_parts(str_ as *const u8, len) };
-    let result = std::str::from_utf8(string);
-    return_error_status_if_false!(env, result.is_ok(), napi_generic_failure);
-    result.unwrap()
+    string
   };
   let result = {
     let env = unsafe { &mut *(env as *mut Env) };
-    v8::String::new(&mut env.scope(), string)
+    v8::String::new_from_utf8(
+      &mut env.scope(),
+      string,
+      v8::NewStringType::Internalized,
+    )
   };
   return_error_status_if_false!(env, result.is_some(), napi_generic_failure);
   Ok(result.unwrap())
@@ -248,10 +248,12 @@ macro_rules! napi_wrap {
       #[inline(always)]
       fn inner $( < $( $x ),* > )? ( $env: & $( $lt )? mut Env , $( $ident : $ty ),* ) -> napi_status $body
 
+      #[cfg(debug_assertions)]
       log::trace!("NAPI ENTER: {}", stringify!($name));
 
       let result = inner( env, $( $ident ),* );
 
+      #[cfg(debug_assertions)]
       log::trace!("NAPI EXIT: {} {}", stringify!($name), result);
 
       if let Some(exception) = try_catch.exception() {
@@ -276,10 +278,12 @@ macro_rules! napi_wrap {
       #[inline(always)]
       fn inner $( < $( $x ),* > )? ( $( $ident : $ty ),* ) -> napi_status $body
 
+      #[cfg(debug_assertions)]
       log::trace!("NAPI ENTER: {}", stringify!($name));
 
       let result = inner( $( $ident ),* );
 
+      #[cfg(debug_assertions)]
       log::trace!("NAPI EXIT: {} {}", stringify!($name), result);
 
       result